# openapi.yaml
openapi: 3.0.3
info:
  title: PromoPing API
  description: |
    API RESTful do PromoPing - Sistema de monitorização de preços de produtos.
    
    A API permite aos utilizadores:
    - Monitorizar preços de produtos em múltiplas lojas online
    - Receber notificações quando os preços atingem valores alvo
    - Gerir produtos, preferências e assinaturas
    - Exportar dados em diferentes formatos
    
    **Autenticação**: A maioria dos endpoints requer autenticação via JWT Bearer Token.
    **Planos**: Algumas funcionalidades estão limitadas por plano (Free, Basic, Standard, Premium).
  version: 2.3.3
  contact:
    name: PromoPing Support
    email: support@promoping.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://127.0.0.1:3000
    description: Servidor de desenvolvimento local
  - url: https://api.promoping.com
    description: Servidor de produção
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Autenticação
      summary: Registar novo utilizador
      description: Cria uma nova conta de utilizador com email e senha
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - email
                - password
              properties:
                nome:
                  type: string
                  example: João Silva
                email:
                  type: string
                  format: email
                  example: joao@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: senha123
                telefone:
                  type: string
                  example: +351912345678
      responses:
        '200':
          description: Conta criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                  codigo:
                    type: string
                    description: Código de verificação (apenas em desenvolvimento)
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Autenticação
      summary: Login com email e senha
      description: Autentica um utilizador e retorna um token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: joao@example.com
                password:
                  type: string
                  format: password
                  example: senha123
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  token:
                    type: string
                    description: JWT token para autenticação
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Credenciais inválidas
        '403':
          description: Email não verificado
...
  /api/produtos:
    get:
      tags:
        - Produtos
      summary: Listar produtos
      description: Retorna todos os produtos do utilizador
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  produtos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Produto'
    post:
      tags:
        - Produtos
      summary: Adicionar produto
      description: Adiciona um novo produto para monitorização (respeitando limite do plano)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - link
                - precoAlvo
              properties:
                nome:
                  type: string
                  example: iPhone 15 Pro
                link:
                  type: string
                  format: uri
                  example: https://www.pcdiga.com/iphone-15-pro
                precoAlvo:
                  type: number
                  minimum: 0
                  example: 999.99
                data:
                  type: string
                  format: date
                  description: Data limite para monitorização
      responses:
        '200':
          description: Produto adicionado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  produto:
                    $ref: '#/components/schemas/Produto'
                  storeInfo:
                    type: object
        '403':
          description: Limite de produtos atingido
...
  /api/notificacoes:
    get:
      tags:
        - Notificações
      summary: Listar notificações
      description: Retorna notificações do utilizador
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de notificações
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  notificacoes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notificacao'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via /api/auth/login
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        telefone:
          type: string
          nullable: true
        fotoPerfil:
          type: string
          nullable: true
        perfilId:
          type: integer
          nullable: true
        DataCriacao:
          type: string
          format: date
...
    Produto:
      type: object
      properties:
        Id:
          type: integer
        Nome:
          type: string
        Link:
          type: string
          format: uri
        PrecoAtual:
          type: number
          nullable: true
        PrecoAlvo:
          type: number
        DataCriacao:
          type: string
          format: date-time
        DataLimite:
          type: string
          format: date
          nullable: true
        Loja:
          type: string
        storeInfo:
          type: object
        Historico:
          type: array
          items:
            type: object
            properties:
              Preco:
                type: number
              Data:
                type: string
    Notificacao:
      type: object
      properties:
        Id:
          type: integer
        Tipo:
          type: string
        Mensagem:
          type: string
        DataEnvio:
          type: string
          format: date-time
        ValorPoupado:
          type: number
          nullable: true
...
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
          description: Mensagem de erro
        message:
          type: string
          description: Mensagem adicional (opcional)